Bug fixes & phi nodes (#57)
* Fix negative numbers

* Fix workflow file for version bump

* chore: bump version

* Added workflow dependencies

* Fix workflow

* Refined emitter logic

* Inherit secrets in workflow

* Introduce phi node. Removed old macro code.

* Changed the version script because we removed main branch.

* Added extra info messages

* Fix bug where self-referential loop causes program to crash

---------

+++ gbf_core/src/bytecode_loader.rs
@@ -321,11 +321,11 @@ impl<R: Read> BytecodeLoader<R> {
             }
             Opcode::ImmByte => {
                 let value = self.reader.read_u8().map_err(BytecodeLoaderError::from)?;
-                Ok(Some((Operand::new_number(value as i32), 1)))
+                Ok(Some((Operand::new_number(value as i8 as i32), 1)))
             }
             Opcode::ImmShort => {
                 let value = self.reader.read_u16().map_err(BytecodeLoaderError::from)?;
-                Ok(Some((Operand::new_number(value as i32), 2)))
+                Ok(Some((Operand::new_number(value as i16 as i32), 2)))
             }
             Opcode::ImmInt => {
                 let value = self.reader.read_u32().map_err(BytecodeLoaderError::from)?;
